add attribute to push more options to base backup
There are other commandline options for backup-push, such as
--cluster-read-rate-limit and --while-offline, et al.

Allowing an operator to specify any or all as a node attribute will
inject these into the cron job's command, in a forma like so:

    node.override[:wal_e][:base_backup][:options] =
    '--cluster-read-rate-limit 5000 --while-offline'

or any such string.

The inital value of `nil` will be represented by a blank, so the command
will have two spaces between `backup-push` and <data directory> if not
provided.